refactor(core): extract runPipeline body, layout packs, segments, bind
Move orchestration into run-pipeline and split layout/bind/scale/geometry
helpers behind stable facades so pipeline modules stay PR-sized and easy
to maintain. Add characterization tests for segment packing and facet
panel placement.
