+++ packages/core/src/pipeline.ts
@@ -36,17 +36,6 @@
  * values with a warning and REFUSE non-positive explicit domains.
  */
 
-import type { SpecInput, PortableSpec } from "@ggsvelte/spec";
-
-import { perfMark, perfMeasure } from "./perf.js";
-
-import type { Advisory, PipelineWarning, RenderModel, RunOptions } from "./pipeline/types.js";
-import { allocatePipelineRunId } from "./pipeline/run-id.js";
-import { setupPipelineRun } from "./pipeline/setup-run.js";
-import { preparePanels } from "./pipeline/prepare-panels.js";
-import { trainPipelineScales } from "./pipeline/train-pipeline-scales.js";
-import { finalizePipelineRun } from "./pipeline/finalize-run.js";
-
 // Re-export the public pipeline contract (import path stability).
 export type {
   Advisory,
@@ -63,60 +52,4 @@ export type {
 } from "./pipeline/public-api.js";
 export { CANVAS_AUTO_THRESHOLD, PipelineError, batchMarkCount } from "./pipeline/public-api.js";
 
-// ---------------------------------------------------------------------------
-// runPipeline
-// ---------------------------------------------------------------------------
-
-export function runPipeline(spec: SpecInput | PortableSpec, options: RunOptions): RenderModel {
-  const runId = allocatePipelineRunId();
-  perfMark("ggsvelte:pipeline:start");
-
-  const warnings: PipelineWarning[] = [];
-  const advisories: Advisory[] = [];
-
-  const { normalized, editionDefaults, theme, flip } = setupPipelineRun(
-    spec,
-    options.editions,
-    warnings,
-  );
-
-  // bind + facet partition + per-panel frames
-  perfMark("ggsvelte:bind:start");
-  const prepared = preparePanels(normalized, options, warnings, advisories);
-  perfMark("ggsvelte:bind:end");
-  perfMeasure("ggsvelte:bind", "ggsvelte:bind:start", "ggsvelte:bind:end");
-
-  // train scales — fixed: union across panels; free: positional domains per
-  // panel; discrete color/fill assignment ALWAYS global (one legend).
-  perfMark("ggsvelte:scales:start");
-  const trained = trainPipelineScales({
-    normalized,
-    options,
-    table: prepared.table,
-    facetPanels: prepared.facetPanels,
-    panelFrames: prepared.panelFrames,
-    freeX: prepared.freeX,
-    freeY: prepared.freeY,
-    editionDefaults,
-    warnings,
-    advisories,
-  });
-  perfMark("ggsvelte:scales:end");
-  perfMeasure("ggsvelte:scales", "ggsvelte:scales:start", "ggsvelte:scales:end");
-
-  const model = finalizePipelineRun({
-    runId,
-    normalized,
-    options,
-    theme,
-    flip,
-    prepared,
-    trained,
-    warnings,
-    advisories,
-  });
-
-  perfMark("ggsvelte:pipeline:end");
-  perfMeasure("ggsvelte:pipeline", "ggsvelte:pipeline:start", "ggsvelte:pipeline:end");
-  return model;
-}
+export { runPipeline } from "./pipeline/run-pipeline.js";
