+++ packages/core/src/pipeline/geometry.ts
@@ -7,23 +7,11 @@ import type { GeometryBatch } from "../scene.js";
 
 import type { LayerFrame, PipelineWarning, ResolvedColorScale } from "./types.js";
 import type { Frame } from "./geometry-shared.js";
-import {
-  areaBatch,
-  glyphsBatch,
-  lineBatch,
-  pointsBatch,
-  rectsBatch,
-  segmentsBatch,
-} from "./geometry-marks.js";
-import { boxplotBatches, errorbarBatch, smoothBatches } from "./geometry-composites.js";
+import { dispatchGeometryBatch } from "./geometry-dispatch.js";
 
 export type { Frame } from "./geometry-shared.js";
 export { flipBatchInPlace } from "./geometry-flip.js";
 export { batchMarkCount } from "./geometry-mark-count.js";
-
-function single(batch: GeometryBatch | null): GeometryBatch[] {
-  return batch === null ? [] : [batch];
-}
 
 export function buildBatch(
   frame: LayerFrame,
@@ -32,28 +20,5 @@ export function buildBatch(
   fill: ResolvedColorScale | null,
   warnings: PipelineWarning[],
 ): GeometryBatch[] {
-  switch (frame.binding.layer.geom) {
-    case "point":
-      return single(pointsBatch(frame, fx, color, warnings));
-    case "line":
-      return single(lineBatch(frame, fx, color, warnings));
-    case "col":
-    case "bar":
-      return single(rectsBatch(frame, fx, fill, warnings));
-    case "area":
-    case "density":
-      return single(areaBatch(frame, fx, fill, warnings));
-    case "rule":
-      return single(segmentsBatch(frame, fx, color, warnings));
-    case "text":
-      return single(glyphsBatch(frame, fx, color, warnings));
-    case "smooth":
-      return smoothBatches(frame, fx, color, fill, warnings);
-    case "boxplot":
-      return boxplotBatches(frame, fx, fill, warnings);
-    case "errorbar":
-      return single(errorbarBatch(frame, fx, color, warnings));
-    default:
-      return [];
-  }
+  return dispatchGeometryBatch(frame, fx, color, fill, warnings);
 }
