+++ packages/core/src/pipeline/panel-layout-chrome.ts
@@ -1,37 +1,16 @@
 /**
  * Panel layout chrome: labs, axis titles (with coord flip), formatters, legends.
  */
-import type { PortableSpec } from "@ggsvelte/spec";
-
-import type { LegendInput, LegendOrder } from "../legend.js";
-import type { PositionScale } from "../scales/train.js";
-import type { ThemeTokens } from "../theme.js";
-
 import { resolvePanelLayoutDisplay } from "./panel-layout-chrome-display.js";
+import type { PanelLayoutChromeInput } from "./panel-layout-chrome-input.js";
 import { resolvePanelLayoutLabs } from "./panel-layout-chrome-labs.js";
 import { resolvePanelLayoutLegends } from "./panel-layout-chrome-legends.js";
 import type { PanelLayoutChrome } from "./panel-layout-chrome-types.js";
-import type { LayerFrame, PipelineWarning, RunOptions } from "./types.js";
 
 export type { PanelLayoutChrome } from "./panel-layout-chrome-types.js";
+export type { PanelLayoutChromeInput } from "./panel-layout-chrome-input.js";
 
-export function resolvePanelLayoutChrome(input: {
-  flip: boolean;
-  freeX: boolean;
-  freeY: boolean;
-  panelScales: readonly { x: PositionScale; y: PositionScale }[];
-  allFrames: readonly LayerFrame[];
-  labs: NonNullable<PortableSpec["labs"]>;
-  scalesConfig: NonNullable<PortableSpec["scales"]>;
-  xScale: PositionScale;
-  yScale: PositionScale;
-  colorLegend: LegendInput | null;
-  fillLegend: LegendInput | null;
-  legendOrder: LegendOrder;
-  theme: ThemeTokens;
-  options: Pick<RunOptions, "width" | "height" | "measureText">;
-  warnings: PipelineWarning[];
-}): PanelLayoutChrome {
+export function resolvePanelLayoutChrome(input: PanelLayoutChromeInput): PanelLayoutChrome {
   const labsChrome = resolvePanelLayoutLabs({
     allFrames: input.allFrames,
     labs: input.labs,
