+++ packages/core/src/pipeline/panel-layout-facet-map.ts
@@ -1,31 +1,13 @@
 /**
  * Map facet panel defs through geometry into placements.
  */
-import type { LayoutTheme, Margins, TickFormatter } from "../layout/layout.js";
-import type { TextMeasurer } from "../layout/measure.js";
-
-import type { FacetPanelDef } from "./facets.js";
 import { placeOneFacetPanel } from "./panel-layout-facet-place-one.js";
-import type { DisplayScalesFn, PanelPlacement } from "./panel-layout-types.js";
+import type { MapFacetPanelPlacementsInput } from "./panel-layout-facet-map-input.js";
+import type { PanelPlacement } from "./panel-layout-types.js";
 
-export function mapFacetPanelPlacements(input: {
-  facetPanels: readonly FacetPanelDef[];
-  freeH: boolean;
-  freeV: boolean;
-  displayScales: DisplayScalesFn;
-  mMax: Margins;
-  panelW: number;
-  panelH: number;
-  colX: number[];
-  rowY: number[];
-  bottomMostRow: number[];
-  hBreaks: readonly (number | string)[] | undefined;
-  vBreaks: readonly (number | string)[] | undefined;
-  formatH: TickFormatter | undefined;
-  formatV: TickFormatter | undefined;
-  measurer: TextMeasurer;
-  layoutTheme: LayoutTheme;
-}): PanelPlacement[] {
+export type { MapFacetPanelPlacementsInput } from "./panel-layout-facet-map-input.js";
+
+export function mapFacetPanelPlacements(input: MapFacetPanelPlacementsInput): PanelPlacement[] {
   const {
     facetPanels,
     freeH,
