+++ packages/core/src/pipeline/panel-layout-placements.ts
@@ -2,8 +2,8 @@
  * Build panel placements for facet grids or a single panel.
  */
 import type { PanelLayoutChrome } from "./panel-layout-chrome.js";
-import { placeFacetPanels } from "./panel-layout-facet.js";
-import { placeSinglePanel } from "./panel-layout-single.js";
+import { placeFacetPanelsFromChrome } from "./panel-layout-placements-facet.js";
+import { placeSinglePanelFromChrome } from "./panel-layout-placements-single.js";
 import type { PanelPlacement } from "./panel-layout-types.js";
 import type { FacetPanelDef } from "./facets.js";
 import type { RunOptions } from "./types.js";
@@ -17,49 +17,8 @@ export function buildPanelPlacements(input: {
   options: Pick<RunOptions, "width">;
 }): PanelPlacement[] {
   const { faceted, nrow, ncol, facetPanels, chrome, options } = input;
-
   if (faceted) {
-    return placeFacetPanels({
-      facetPanels,
-      nrow,
-      ncol,
-      freeH: chrome.freeH,
-      freeV: chrome.freeV,
-      outerLeftTitle: chrome.vTitle,
-      outerBottomTitle: chrome.hTitle,
-      axisTitleBand: chrome.axisTitleBand,
-      legendWidth: chrome.legendBlock.width,
-      optionsWidth: options.width,
-      layoutHeight: chrome.layoutHeight,
-      topBand: chrome.topBand,
-      displayScales: chrome.displayScales,
-      hBreaks: chrome.hBreaks,
-      vBreaks: chrome.vBreaks,
-      formatH: chrome.formatH,
-      formatV: chrome.formatV,
-      measurer: chrome.measurer,
-      layoutTheme: chrome.layoutTheme,
-    });
+    return placeFacetPanelsFromChrome({ nrow, ncol, facetPanels, chrome, options });
   }
-
-  const { h, v } = chrome.displayScales(0);
-  return [
-    placeSinglePanel({
-      h,
-      v,
-      hTitle: chrome.hTitle,
-      vTitle: chrome.vTitle,
-      axisTitleBand: chrome.axisTitleBand,
-      legendWidth: chrome.legendBlock.width,
-      optionsWidth: options.width,
-      layoutHeight: chrome.layoutHeight,
-      topBand: chrome.topBand,
-      hBreaks: chrome.hBreaks,
-      vBreaks: chrome.vBreaks,
-      formatH: chrome.formatH,
-      formatV: chrome.formatV,
-      measurer: chrome.measurer,
-      layoutTheme: chrome.layoutTheme,
-    }),
-  ];
+  return [placeSinglePanelFromChrome(chrome, options)];
 }
