+++ packages/core/src/pipeline/scale-axis-collect-acc.ts
@@ -1,0 +1,14 @@
+/**
+ * Shared accumulator for axis evidence collection.
+ */
+import type { CellValue } from "../table.js";
+
+export interface AxisCollectAcc {
+  columns: (readonly CellValue[])[];
+  numeric: Float64Array[];
+  anyDiscrete: boolean;
+  allTemporal: boolean;
+  sawContinuousEvidence: boolean;
+  barMeasure: boolean;
+  typeParts: Set<string>;
+}
