+++ packages/core/src/pipeline/train-pipeline-scales.ts
@@ -1,31 +1,15 @@
 /**
  * Train fixed/free positional scales and global color/fill scales for a run.
  */
-import type { PortableSpec } from "@ggsvelte/spec";
-
-import type { EditionDefaults } from "../editions.js";
-import type { ColumnTable } from "../table.js";
-
-import type { FacetPanelDef } from "./facets.js";
 import { trainPipelineColorScales } from "./train-pipeline-scales-color.js";
+import type { TrainPipelineScalesInput } from "./train-pipeline-scales-input.js";
 import { trainPipelinePositionScales } from "./train-pipeline-scales-position.js";
 import type { TrainedPipelineScales } from "./train-pipeline-scales-types.js";
-import type { Advisory, LayerFrame, PipelineWarning, RunOptions } from "./types.js";
 
 export type { TrainedPipelineScales } from "./train-pipeline-scales-types.js";
+export type { TrainPipelineScalesInput } from "./train-pipeline-scales-input.js";
 
-export function trainPipelineScales(input: {
-  normalized: PortableSpec;
-  options: RunOptions;
-  table: ColumnTable;
-  facetPanels: readonly FacetPanelDef[];
-  panelFrames: readonly (readonly LayerFrame[])[];
-  freeX: boolean;
-  freeY: boolean;
-  editionDefaults: EditionDefaults;
-  warnings: PipelineWarning[];
-  advisories: Advisory[];
-}): TrainedPipelineScales {
+export function trainPipelineScales(input: TrainPipelineScalesInput): TrainedPipelineScales {
   const {
     normalized,
     options,
